requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
black==19.10b0
bottle==0.12.18
click==7.1.2
Django==3.1.2
djangorestframework==3.12.1
environs==8.0.0
funcy==1.14
gunicorn==20.0.4
honcho==1.0.1
importlib-metadata==2.0.0
iniconfig==1.1.1
marshmallow==3.8.0
packaging==20.4
pathspec==0.8.0
peewee==3.13.3
pkg-resources==0.0.0
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pyparsing==2.4.7
pytest==6.1.1
python-dotenv==0.14.0
pytz==2020.1
regex==2020.7.14
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
zipp==3.4.0
